Partial Commit for backup:     The changes are made with the introduction of roles app that is responsible for checking user permission of a role associated with user     The role is implemented partially for the test purpose     This commit feature is not for final use and is only for backup purpose
    changes :
      1) able to add user as dev in the command line
      2) Otp login is now done with the phonenumber instead of username
      3) api in user will check for permissions of the logged in user

    Todo :
      1) Implement role based access to all models
      2) Enable security featurnes
      3) Frontend integration for Roles
      4) Enable site admin to administer every models
      5) Implement role creation API

+++ user/views.py
@@ -16,18 +16,26 @@ from .serializer import UserRegistrationSerializer, LoginSerializer
 from .otp import OTPLIST
 from rest_framework_simplejwt.tokens import RefreshToken
 from captcha.helpers import captcha_image_url
+from roles.views import is_role_capable_of
+from roles.models import Role
 
 def get_captcha(request):
     hashkey = CaptchaStore.generate_key()
     imageurl = captcha_image_url(hashkey)
 
-    response = JsonResponse({
+    send_response = JsonResponse({
         'key': hashkey,
         'image_url': imageurl
     })
-    return response  
+    return send_response
 
-# UserAPI class handles user registration, fetching user data, updating user details, and deleting user accounts.
+
+'''
+ UserAPI class handles user registration, fetching user data,
+ updating user details, and deleting user accounts.
+'''
+
+
 class UserAPI(APIView):
 
     # POST method for user registration
@@ -43,6 +51,13 @@ class UserAPI(APIView):
 
     # GET method to retrieve user data by username or the logged-in user
     def get(self, request, username=None, *args, **kwargs):
+
+        if is_role_capable_of(request=request,
+                              operation=Role.READ,
+                              model='user') is False:
+
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         if username:
             try:
                 user = CustomUser.objects.get(username=username)
@@ -85,6 +100,12 @@ class UserAPI(APIView):
 
     # PUT method to update user details
     def put(self, request, username, *args, **kwargs):
+        if is_role_capable_of(request=request,
+                              operation=Role.READ_WRITE,
+                              model='user') is False:
+
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         try:
             data = json.loads(request.body)
             success = update_user_details(
@@ -106,6 +127,12 @@ class UserAPI(APIView):
 
     # DELETE method to remove a user by username
     def delete(self, request, username, *args, **kwargs):
+        if is_role_capable_of(request=request,
+                              operation=Role.READ_WRITE,
+                              model='user') is False:
+
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         try:
             delete_user_by_username(username)
             return JsonResponse({'message': 'User deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
@@ -113,13 +140,16 @@ class UserAPI(APIView):
             return JsonResponse({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
 # LoginAPI handles user login functionality via JWT.
+
+
 class LoginAPI(APIView):
     serializer_class = LoginSerializer
 
     # POST method for logging in the user
     def post(self, request):
         serializer = self.serializer_class(data=request.data)
-        serializer.is_valid(raise_exception=True)  # Automatically raises ValidationError if invalid
+        # Automatically raises ValidationError if invalid
+        serializer.is_valid(raise_exception=True)
         validated_data = serializer.validated_data  # Extract validated data
 
         response_data = {
@@ -135,6 +165,8 @@ class LoginAPI(APIView):
         return Response(response_data, status=status.HTTP_200_OK)
 
 # LogoutAPI handles the user logout by invalidating the refresh token.
+
+
 class LogoutAPI(APIView):
 
     # POST method for logging out by blacklisting the refresh token
@@ -152,28 +184,34 @@ class LogoutAPI(APIView):
         except Exception as e:
             return Response({"error": str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
+
 # OTP Handling: Static OTP list to manage OTP requests and verification.
 static_otp_list = OTPLIST()
 
 # send_otp_API generates and sends OTP for a given username
+
+
 def send_otp_API(request):
     if request.method == 'POST':
-        username = request.POST.get('username')
+        phonenumber = request.POST.get('phonenumber')
 
-        if not username:
-            return JsonResponse({'error': 'username is required for otp login'})
+        if not phonenumber:
+            return JsonResponse({'error': 'phonenumber is required for otp login'})
 
         try:
-            user = CustomUser.objects.get(username=username)  # Ensure the user exists
+            user = CustomUser.objects.get(phonenumber=phonenumber)
+            # Ensure the user exists
         except User.DoesNotExist:
-            return JsonResponse({'error': 'user with this username does not exist'})
+            return JsonResponse({'error': 'user with this phonenumber does not exist'})
 
         # Cleanup and check expired OTPs before generating a new one
         if len(static_otp_list.otplist) > 0:
             static_otp_list.check_time_and_mark()
             static_otp_list.cleanup()
 
-        otp = static_otp_list.get_new_otp(in_username=username)  # Generate new OTP
+        otp = static_otp_list.get_new_otp(
+            in_username=phonenumber)  # Generate new OTP
+
         print(f"OTP: {otp.otp}")  # For debugging purpose
 
         return JsonResponse({'index': otp.index})
@@ -182,6 +220,8 @@ def send_otp_API(request):
         return JsonResponse({'error': 'Invalid request sent'})
 
 # verify_otp_API checks and verifies the OTP entered by the user
+
+
 def verify_otp_API(request):
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -211,7 +251,7 @@ def verify_otp_API(request):
             }
 
             return JsonResponse(response_data, status=status.HTTP_200_OK)
-        
+
         else:
             return JsonResponse({'error': 'wrong otp'})
